Added support to explicitly send standalone LTRS and FIGS characters to loop
This is particularly useful for ensuring a printer is in the correct shift when starting a new session with it. Use { to shift into LTRS and } to shift into FIGS. Note that the machine will stay in that shift despite sending characters of the opposite shift until after a space is sent.